+++ CHANGELOG.md
@@ -21,3 +21,4 @@
 - [2026-06-17T13:00:05.882Z] feat: implement Tracker tab
 - [2026-06-17T18:30:42.128Z] feat: add dashboard components
 - [2026-06-17T19:15:48.753Z] style: improve dashboard layout
+- [2026-06-17T22:52:33.110Z] feat: integrate Dashboard tab